[greenplumwriter] Optimize write performance
