+++ plugin/writer/greenplumwriter/src/main/java/com/wgzhao/addax/plugin/writer/greenplumwriter/GPConstant.java
@@ -19,10 +19,11 @@
 
 package com.wgzhao.addax.plugin.writer.greenplumwriter;
 
-public class GPConstant
-{
-    public static final int TIME_OUT_MS = 5000;
+import com.wgzhao.addax.common.base.Constant;
 
+public class GPConstant
+        extends Constant
+{
     // 字段分隔符
     public static final char DELIMITER = '\u0001';
 
@@ -35,15 +36,4 @@ public class GPConstant
 
     // 因为 GP DB 服务端 对 COPY FROM 的 CSV 格式做了这样的限制，如果单个元组大于4 MB，只能使用 insert
     public static final int MAX_CSV_SIZE = 4194304;
-
-    //  线程异步队列大小，增大此参数增加内存消耗，提升性能
-    public static final int COPY_QUEUE_SIZE = 1000;
-
-    // 用于进行格式化数据的线程数
-    public static final int NUM_COPY_PROCESSOR = 4;
-
-    // 写入数据库的并发数
-    public static final int NUM_COPY_WRITER = 1;
-
-
 }
