+++ plugin/writer/greenplumwriter/src/main/java/com/wgzhao/addax/plugin/writer/greenplumwriter/GreenplumWriter.java
@@ -100,8 +100,7 @@ public class GreenplumWriter
 
         public void startWrite(RecordReceiver recordReceiver)
         {
-            this.commonRdbmsWriterTask.startWrite(recordReceiver, this.writerSliceConfig,
-                    super.getTaskPluginCollector());
+            this.commonRdbmsWriterTask.startWrite(recordReceiver, this.writerSliceConfig, super.getTaskPluginCollector());
         }
 
         @Override
