+++ plugin/writer/greenplumwriter/src/main/resources/plugin_job_template.json
@@ -3,12 +3,10 @@
   "parameter": {
     "username": "",
     "password": "",
-    "queueSize": 1000,
-    "numProc": 4,
-    "numWriter": 1,
     "column": [
       "*"
     ],
+    "batchSize": "1024",
     "preSql": [],
     "connection": [
       {
